refactor: separate charactersState from mainState
